Add files via upload
This spreadsheet contains the data that were used as the foundation to implement the synthetic control method. They include data related to the Consumer Price Index of Lithuania and other countries in the control group (those similar to Lithuania but that did not adopt the Euro as their currency). Additionally, the GDP of these countries is also included.